server/requirements.txt
uwsgi
django-phonenumber-field
Django
djangorestframework
psycopg2
